requirements.txt
# streamlit
python-dotenv
# streamlit_option_menu
# streamlit_lottie
load_dotenv
crewai
crewai_tools
python-docx
textwrap3
# langchain_google_genai
# langchain-openai
langchain_groq
opencv-python
numpy
unstructured
streamlit
streamlit_lottie
-e .